pyproject.toml
[project]
name = "telegram-stt-bot"
version = "0.1.0"
description = "Telegram bot for speech-to-text conversion"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=24.1.0",
    "aiohttp>=3.12.15",
    "fastapi>=0.116.1",
    "loguru>=0.7.3",
    "openai-whisper>=20250625",
    "python-multipart>=0.0.20",
    "python-telegram-bot>=22.4",
    "torch>=2.8.0",
    "uvicorn>=0.35.0",
]

[dependency-groups]
dev = [
    "mypy>=1.18.1",
    "notebook>=7.4.5",
    "pytest>=8.4.2",
    "ruff>=0.13.0",
]
